task_manager/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
gitdb==4.0.12
GitPython==3.1.44
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
smmap==5.0.2
Werkzeug==3.1.3
